+++ application/module/users/interfaces/users.php
@@ -1,16 +1,16 @@
 <?php
-namespace Nibiru\Messages;
+namespace Nibiru\Module\Users\Interfaces;
 /**
  * @desc this file is for the autoloader to function properly,
- * you might as well use it as the primary user trait for your
+ * you might as well use it as the primary user interface for your
  * application
  * Created by PhpStorm.
  * User: kasdorf
  * Date: 28.08.18
- * Time: 11:24
+ * Time: 11:21
  */
 
-trait Users
+interface Users
 {
 
 }
