+++ application/settings/config/settings.development.ini
@@ -26,9 +26,9 @@ register.text           = "PATHTOYOURHTMLTEMPLATE"
 pageurl                =   "YOURPAGEURLWITHPROTOCOL"
 navigation             =   "navigation.json"
 dbmodel                =   "/../../application/model/"
-module                 =   "/../../application/module/"
-interfaces             =   "/../../application/interfaces/"
-trait                  =   "/../../application/trait/"
+module                 =   "/../../application/module/[NAME]"
+interfaces             =   "/../../application/module/[NAME]/interfaces/"
+traits                 =   "/../../application/module/[NAME]/traits/"
 entriesperpage         =   4
 background.img[]       =   "public/img/nibiru3.jpg"
 smarty.css[]           =   "public/css/v3/roboto.css"
